[minor/bug] wrong cast... (float -> double)
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@4672 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/Misc/moneyprint/moneyprint.c
@@ -33,7 +33,7 @@ int main( int argc , char *argv[]){
     int cnt;
     char *ptr;
     char *sign;
-    float n,t2;
+    double n,t2;
     int t1;
     int t3;
     cnt = 0;
